Updates to CPI system
