+++ Cargo.toml
@@ -27,9 +27,10 @@ rayon = "1.10.0"
 run_script = "0.11.0"
 indicatif = "0.17.11"
 ssh2 = { version = "0.9.0" }
-libloading = { version = "0.8.6" }
+libloading = "0.8.6"
 lib_cpi = "0.1.0"
 dirs = "6.0.0"
+comfy-table = "7.1.4"
 
 [profile.dev]
 codegen-units = 32
