+++ src/cpis/prelude.rs
@@ -1,0 +1,3 @@
+pub use super::error::CpiError;
+
+pub type Result<T> = std::result::Result<T, CpiError>;
